+++ src/inspect_ai/_view/www/src/@types/log.d.ts
@@ -194,23 +194,51 @@ export type Content =
       | ContentImage
       | ContentAudio
       | ContentVideo
+      | ContentData
     )[];
 export type Type3 = "text";
 export type Text = string;
 export type Refusal = boolean | null;
-export type Type4 = "reasoning";
+export type Citations =
+  | (ContentCitation | DocumentCitation | UrlCitation)[]
+  | null;
+export type CitedText = string | [number, number] | null;
+export type Title = string | null;
+export type Internal = {
+  [k: string]: JsonValue;
+} | null;
+export type JsonValue = unknown;
+export type Type4 = "content";
+export type CitedText1 = string | [number, number] | null;
+export type Title1 = string | null;
+export type Internal1 = {
+  [k: string]: JsonValue;
+} | null;
+export type Type5 = "document";
+export type Type6 = "block" | "page" | "char";
+export type StartIndex = number;
+export type EndIndex = number;
+export type CitedText2 = string | [number, number] | null;
+export type Title2 = string | null;
+export type Internal2 = {
+  [k: string]: JsonValue;
+} | null;
+export type Type7 = "url";
+export type Url = string;
+export type Type8 = "reasoning";
 export type Reasoning = string;
 export type Signature = string | null;
 export type Redacted = boolean;
-export type Type5 = "image";
+export type Type9 = "image";
 export type Image = string;
 export type Detail = "auto" | "low" | "high";
-export type Type6 = "audio";
+export type Type10 = "audio";
 export type Audio = string;
 export type Format1 = "wav" | "mp3";
-export type Type7 = "video";
+export type Type11 = "video";
 export type Video = string;
 export type Format2 = "mp4" | "mpeg" | "mov";
+export type Type12 = "data";
 export type Source = ("input" | "generate") | null;
 export type Role = "system";
 export type Id2 = string | null;
@@ -222,6 +250,7 @@ export type Content1 =
       | ContentImage
       | ContentAudio
       | ContentVideo
+      | ContentData
     )[];
 export type Source1 = ("input" | "generate") | null;
 export type Role1 = "user";
@@ -235,6 +264,7 @@ export type Content2 =
       | ContentImage
       | ContentAudio
       | ContentVideo
+      | ContentData
     )[];
 export type Source2 = ("input" | "generate") | null;
 export type Role2 = "assistant";
@@ -242,10 +272,10 @@ export type ToolCalls = ToolCall[] | null;
 export type Id4 = string;
 export type Function = string;
 export type ParseError = string | null;
-export type Title = string | null;
+export type Title3 = string | null;
 export type Format3 = "text" | "markdown";
 export type Content3 = string;
-export type Type8 = string | null;
+export type Type13 = string | null;
 export type Model2 = string | null;
 export type Id5 = string | null;
 export type Content4 =
@@ -256,12 +286,13 @@ export type Content4 =
       | ContentImage
       | ContentAudio
       | ContentVideo
+      | ContentData
     )[];
 export type Source3 = ("input" | "generate") | null;
 export type Role3 = "tool";
 export type ToolCallId1 = string | null;
 export type Function1 = string | null;
-export type Type9 =
+export type Type14 =
   | "parsing"
   | "timeout"
   | "unicode_decode"
@@ -344,13 +375,12 @@ export type Files1 = {
   [k: string]: string;
 } | null;
 export type Setup1 = string | null;
-export type JsonValue = unknown;
 export type SpanId1 = string | null;
 export type Timestamp1 = string;
 export type WorkingStart1 = number;
 export type Pending1 = boolean | null;
 export type Event1 = "sample_limit";
-export type Type10 =
+export type Type15 =
   | "message"
   | "time"
   | "working"
@@ -404,7 +434,7 @@ export type Input3 = (
 )[];
 export type Name8 = string;
 export type Description2 = string;
-export type Type11 = "object";
+export type Type16 = "object";
 export type Required1 = string[];
 export type Additionalproperties1 = boolean;
 export type Options3 = {
@@ -424,7 +454,7 @@ export type Timestamp6 = string;
 export type WorkingStart6 = number;
 export type Pending6 = boolean | null;
 export type Event6 = "tool";
-export type Type12 = "function";
+export type Type17 = "function";
 export type Id7 = string;
 export type Function2 = string;
 export type Result1 =
@@ -436,12 +466,14 @@ export type Result1 =
   | ContentImage
   | ContentAudio
   | ContentVideo
+  | ContentData
   | (
       | ContentText
       | ContentReasoning
       | ContentImage
       | ContentAudio
       | ContentVideo
+      | ContentData
     )[];
 export type Truncated = [unknown, unknown] | null;
 export type SpanId7 = string | null;
@@ -510,7 +542,7 @@ export type Pending13 = boolean | null;
 export type Event13 = "span_begin";
 export type Id8 = string;
 export type ParentId = string | null;
-export type Type13 = string | null;
+export type Type18 = string | null;
 export type Name11 = string;
 export type SpanId14 = string | null;
 export type Timestamp14 = string;
@@ -524,7 +556,7 @@ export type WorkingStart15 = number;
 export type Pending15 = boolean | null;
 export type Event15 = "step";
 export type Action1 = "begin" | "end";
-export type Type14 = string | null;
+export type Type19 = string | null;
 export type Name12 = string;
 export type SpanId16 = string | null;
 export type Timestamp16 = string;
@@ -532,7 +564,7 @@ export type WorkingStart16 = number;
 export type Pending16 = boolean | null;
 export type Event16 = "subtask";
 export type Name13 = string;
-export type Type15 = string | null;
+export type Type20 = string | null;
 export type Events2 = (
   | SampleInitEvent
   | SampleLimitEvent
@@ -600,7 +632,7 @@ export type TotalTime = number | null;
 export type WorkingTime3 = number | null;
 export type Uuid = string | null;
 export type ErrorRetries = EvalError[] | null;
-export type Type16 =
+export type Type21 =
   | "context"
   | "time"
   | "working"
@@ -1014,6 +1046,44 @@ export interface ContentText {
   type: Type3;
   text: Text;
   refusal: Refusal;
+  citations: Citations;
+}
+/**
+ * A generic content citation.
+ */
+export interface ContentCitation {
+  cited_text: CitedText;
+  title: Title;
+  internal: Internal;
+  type: Type4;
+}
+/**
+ * A citation that refers to a page range in a document.
+ */
+export interface DocumentCitation {
+  cited_text: CitedText1;
+  title: Title1;
+  internal: Internal1;
+  type: Type5;
+  range: DocumentRange | null;
+}
+/**
+ * A range specifying a section of a document.
+ */
+export interface DocumentRange {
+  type: Type6;
+  start_index: StartIndex;
+  end_index: EndIndex;
+}
+/**
+ * A citation that refers to a URL.
+ */
+export interface UrlCitation {
+  cited_text: CitedText2;
+  title: Title2;
+  internal: Internal2;
+  type: Type7;
+  url: Url;
 }
 /**
  * Reasoning content.
@@ -1022,7 +1092,7 @@ export interface ContentText {
  */
 export interface ContentReasoning {
   internal: unknown;
-  type: Type4;
+  type: Type8;
   reasoning: Reasoning;
   signature: Signature;
   redacted: Redacted;
@@ -1032,7 +1102,7 @@ export interface ContentReasoning {
  */
 export interface ContentImage {
   internal: unknown;
-  type: Type5;
+  type: Type9;
   image: Image;
   detail: Detail;
 }
@@ -1041,7 +1111,7 @@ export interface ContentImage {
  */
 export interface ContentAudio {
   internal: unknown;
-  type: Type6;
+  type: Type10;
   audio: Audio;
   format: Format1;
 }
@@ -1050,9 +1120,20 @@ export interface ContentAudio {
  */
 export interface ContentVideo {
   internal: unknown;
-  type: Type7;
+  type: Type11;
   video: Video;
   format: Format2;
+}
+/**
+ * Model internal.
+ */
+export interface ContentData {
+  internal: unknown;
+  type: Type12;
+  data: Data;
+}
+export interface Data {
+  [k: string]: JsonValue;
 }
 /**
  * User chat message.
@@ -1084,7 +1165,7 @@ export interface ToolCall {
   internal: unknown;
   parse_error: ParseError;
   view: ToolCallContent | null;
-  type: Type8;
+  type: Type13;
 }
 export interface Arguments {
   [k: string]: unknown;
@@ -1093,7 +1174,7 @@ export interface Arguments {
  * Content to include in tool call view.
  */
 export interface ToolCallContent {
-  title: Title;
+  title: Title3;
   format: Format3;
   content: Content3;
 }
@@ -1111,7 +1192,7 @@ export interface ChatMessageTool {
   error: ToolCallError | null;
 }
 export interface ToolCallError {
-  type: Type9;
+  type: Type14;
   message: Message1;
 }
 /**
@@ -1205,7 +1286,7 @@ export interface SampleLimitEvent {
   working_start: WorkingStart1;
   pending: Pending1;
   event: Event1;
-  type: Type10;
+  type: Type15;
   message: Message2;
   limit: Limit1;
 }
@@ -1322,7 +1403,7 @@ export interface ToolInfo {
  * Description of tool parameters object in JSON Schema format.
  */
 export interface ToolParams {
-  type: Type11;
+  type: Type16;
   properties: Properties1;
   required: Required1;
   additionalProperties: Additionalproperties1;
@@ -1356,7 +1437,7 @@ export interface ToolEvent {
   working_start: WorkingStart6;
   pending: Pending6;
   event: Event6;
-  type: Type12;
+  type: Type17;
   id: Id7;
   function: Function2;
   arguments: Arguments1;
@@ -1486,7 +1567,7 @@ export interface SpanBeginEvent {
   event: Event13;
   id: Id8;
   parent_id: ParentId;
-  type: Type13;
+  type: Type18;
   name: Name11;
 }
 /**
@@ -1510,7 +1591,7 @@ export interface StepEvent {
   pending: Pending15;
   event: Event15;
   action: Action1;
-  type: Type14;
+  type: Type19;
   name: Name12;
 }
 /**
@@ -1523,7 +1604,7 @@ export interface SubtaskEvent {
   pending: Pending16;
   event: Event16;
   name: Name13;
-  type: Type15;
+  type: Type20;
   input: Input5;
   result: Result2;
   events: Events2;
@@ -1546,7 +1627,7 @@ export interface Attachments {
  * Limit encountered by sample.
  */
 export interface EvalSampleLimit {
-  type: Type16;
+  type: Type21;
   limit: Limit2;
 }
 /**