+++ src/inspect_ai/model/_call_tools.py
@@ -36,6 +36,7 @@ from pydantic import BaseModel
 from inspect_ai._util.content import (
     Content,
     ContentAudio,
+    ContentData,
     ContentImage,
     ContentText,
     ContentVideo,
@@ -188,13 +189,19 @@ async def execute_tools(
             # types to string as that is what the model APIs accept
             truncated: tuple[int, int] | None = None
             if isinstance(
-                result, ContentText | ContentImage | ContentAudio | ContentVideo
+                result,
+                ContentText | ContentImage | ContentAudio | ContentVideo | ContentData,
             ):
                 content: str | list[Content] = [result]
             elif isinstance(result, list) and (
                 len(result) == 0
                 or isinstance(
-                    result[0], ContentText | ContentImage | ContentAudio | ContentVideo
+                    result[0],
+                    ContentText
+                    | ContentImage
+                    | ContentAudio
+                    | ContentVideo
+                    | ContentData,
                 )
             ):
                 content = result