+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+- Anthropic: Support for binding to internal `web search()` tool.
+- Content: Added `ContentData` for model specific content blocks.
+- Citations: Added `Citation` suite of types and included citations in `ContentText` (supported for OpenAI and Anthropic models).
+
 ## v0.3.103 (06 June 2025)
 
 - Eval set: Do not read full eval logs into memory at task completion.